fix(config): BigDecimal for rate, real filter chain test, CommissionRateResponse in dto package
- Change Commission.rate from double to BigDecimal in DonyConfigProperties (financial precision)
- Change CommissionRateResponse to use BigDecimal, moved to config/dto/ subpackage
- Rewrite ConfigControllerIT with explicit Javadoc explaining why @SpringBootTest +
  @AutoConfigureMockMvc runs the real filter chain (not a false positive)
- Add Javadoc on DonyConfigProperties noting legacy flat-key usage in PaymentService

+++ src/main/java/com/dony/api/config/DonyConfigProperties.java
@@ -2,9 +2,16 @@ package com.dony.api.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
+import java.math.BigDecimal;
+
+/**
+ * Dony application configuration properties (prefix "dony").
+ * Note: dony.stripe.* and dony.commission-rate (legacy flat key) are
+ * intentionally consumed via @Value in PaymentService for now.
+ */
 @ConfigurationProperties(prefix = "dony")
 public record DonyConfigProperties(
     Commission commission
 ) {
-    public record Commission(double rate) {}
+    public record Commission(BigDecimal rate) {}
 }
